requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-postgres
requests
psycopg2-binary
python-dotenv
pandas